fix: create resto without owner, add lat/lon on get restos

+++ src/main/java/com/roulette/resto/data/resto/entity/Restaurant.java
@@ -24,8 +24,8 @@ public class Restaurant {
 	Set<String> foodTypes;
 	Set<String> labels;
 	String address;
-	BigDecimal longitude;
-	BigDecimal latitude;
+	double longitude;
+	double latitude;
 	List<BusinessHour> businessHours;
 	Date creationDate;
 	Account owner;
